Fix nacl warnings (#118)
* firt commits to fix nacl warnings

* some fixes

* fixed format issues

* almost done with nacl_syscall_common.c

* more fixes

* got rid of unreferenced variable warning

* almost there

* more fixes

* final fixes

* final fix

* fixing comment

* fix Nick's comments

* last fix

* found another extra space

* fixes

* new line fix

* fix ioctl

* another fix

+++ src/trusted/service_runtime/nacl_copy.c
@@ -68,7 +68,7 @@ int NaClCopyInFromUserZStr(struct NaClApp *nap,
   }
 
   uintptr_t check_addr = src_usr_addr;
-  int bytes_copied = 0;
+  size_t bytes_copied = 0;
 
   /*
    * We need to handle NULL terminated strings that possibly could overlap VMMap entries,
